blacklist

+++ authentication/middleware.py
@@ -1,5 +1,5 @@
 from django.utils import timezone
-from .models import Session
+from .models import Session, BlackList
 from rest_framework.exceptions import AuthenticationFailed
 from rest_framework.authentication import BaseAuthentication
 
@@ -21,11 +21,17 @@ class CustomSessionAuthentication(BaseAuthentication):
 
         try:
             session = Session.objects.get(session_token=session_token)
-            if session.expires_at > timezone.now():
-                custom_user = CustomUser(supervisor=session.supervisor, customer=session.customer)
-                return (custom_user, None)
-            else:
-                session.delete()
-                raise AuthenticationFailed('Session expired')
         except Session.DoesNotExist:
             raise AuthenticationFailed('Invalid session token')
+
+        blacklisted_ip = BlackList.objects.filter(ip=session.ip).first()
+        
+        if blacklisted_ip:
+            raise AuthenticationFailed('Your IP is blacklisted')
+        
+        if session.expires_at > timezone.now():
+            custom_user = CustomUser(supervisor=session.supervisor, customer=session.customer)
+            return (custom_user, None)
+        else:
+            session.delete()
+            raise AuthenticationFailed('Session expired')
